+++ CHANGELOG.md
@@ -5,3 +5,13 @@
  - Initial package release based on a cleaned up version of my Gist originally published here:
  
     https://gist.github.com/Preston-Landers/267391562bc96959eb41
+
+ Compared to the original version of the Gist, this new version:
+ 
+  * Has been updated for modern Python 3 (and 2.7)
+  
+  * Correctly handles complex script command lines, e.g. quoted spaces in parameters.
+  
+  * Provides a decorator for your script's main function for easier usage.
+
+  * It has some basic tests.